requirements.txt
cryptography>=2.8,<3.0
PyJWT>=1.7.1,<2.0
requests>=2.23.0,<3.0
cachetools